+++ almost_make/utils/shellUtil/__init__.py
@@ -1,3 +1,3 @@
 #!/usr/bin/python3
 
-__all__ = [ "escapeParser.py", "runner.py", "shellUtil.py", "interactiveShell.py", "globber.py" ]
+__all__ = [ "escapeParser", "runner", "shellUtil", "interactiveShell", "globber" ]