+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 0.0.16
+ * Built-in `grep` and `cat` commands.
+ * Globbing bug-fixes.
+ * `.include`, `include`, `sinclude`, and `-include` directives.
+    * `.include` is not fully compatible with BSDMake.
+
 ## 0.0.15
  * Built-in `ls`
     * Doesn't use color when piped (assuming pipes not sent to system)